Add method to hash password using bcrypt
